virkar bara í fyrsta iteration

+++ c++/MCTS/MCTSAgent.h
@@ -8,7 +8,7 @@ using namespace std;
 
 struct IterationValue {
   int column;
-  array<int, COLUMNS> q_values;
+  array<double, COLUMNS> q_values;
   array<int, COLUMNS> n_values;
   array<double, COLUMNS> policy;
 };
